use common validation logic
