+++ plugins/com.google.cloud.tools.eclipse.ui.util/META-INF/MANIFEST.MF
@@ -12,6 +12,7 @@ Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
  com.google.cloud.tools.eclipse.ui.util,
  com.google.cloud.tools.eclipse.util,
  com.google.cloud.tools.eclipse.util.status,
+ com.google.cloud.tools.project;version="0.1.9",
  com.google.common.annotations;version="15.0.0",
  org.eclipse.core.commands,
  org.eclipse.core.databinding.conversion,
